update move and combat angular

+++ UcmEmpireAngular/src/app/board/_components/board/board.component.ts
@@ -24,14 +24,28 @@ export class BoardComponent implements OnInit {
     this.refresh();
   }
 
-  move(cell) {
+  onClick(cell) {
     if (this.first == null)
-      this.first = cell;
+      this.first = this.board[cell.rowId][cell.id].content == null ? null : this.board[cell.rowId][cell.id].special ? null : cell;
     else {
-      this.boardService.move([this.first, cell]).subscribe(() => {
-        sessionStorage.clear();
-        this.refresh();
-      });
+      if (this.board[cell.rowId][cell.id].content == null) {
+        this.boardService.move([this.first, cell]).subscribe(() => {
+          sessionStorage.clear();
+          this.refresh();
+        });
+        // attack
+      } else if (this.board[this.first.rowId][this.first.id].content.damage && !this.board[cell.rowId][cell.id].special) {
+        if (this.board[cell.rowId][cell.id].content.idUser != this.board[this.first.rowId][this.first.id].content.idUser) {
+          this.board[cell.rowId][cell.id].content.hp -= this.board[this.first.rowId][this.first.id].content.damage;
+          if (this.board[cell.rowId][cell.id].content.hp <= 0) {
+            this.boardService.deathEntity(cell).subscribe(() => {
+              sessionStorage.clear();
+              this.refresh();
+            });
+          } else this.first = null;
+        }
+      } else
+        this.first = null;
     }
   }
 
@@ -39,16 +53,37 @@ export class BoardComponent implements OnInit {
     this.rows = [];
     this.first = null;
     this.boardService.getBoard().subscribe(board => {
+      this.board = board;
       this.dimension = board.length;
       for (let i = 0; i < this.dimension; i++) {
         for (let j = 0; j < this.dimension; j++) {
-          sessionStorage.setItem('' + i + j, board[i][j].content != null ? board[i][j].special ? board[i][j].content.resourceName : 'SOLDAT' : null);
+          let content;
+
+          if (board[i][j].content == null)
+            content = null;
+          else {
+            if (board[i][j].special)
+              content = board[i][j].content.resourceName;
+            else {
+              if (board[i][j].content.damage) {
+                content = 'SOLDAT';
+              } else {
+                content = 'FARMER'
+              }
+              if (board[i][j].content != null && board[i][j].content.idUser == 0)
+                content += '_BLUE';
+              else
+                content += '_RED';
+            }
+          }
+          sessionStorage.setItem('' + i + j, content);
         }
       }
       for (let i = 0; i < this.dimension; i++) {
         this.rows.push({
           dimension: this.dimension,
-          id: i
+          id: i,
+          row: board[i]
         });
       }
     });
